fix: Change dividend income calculations
